NEWS: PB Lang now part of Highlight bundle
Changed Highlight README's to include news that since v3.33 Highlight bundles the PureBASIC language file of this repo.

NOTE: Also the PB theme was to be bundled, but it was mistakenly left out; I mention this in README file, and have added due notes and instructions for manual installation until it gets fixed.